fetching from hyperliquid

+++ src/lib.rs
@@ -7,14 +7,15 @@ use solana_program::{
     pubkey::Pubkey,
 };
 use serde::{Serialize, Deserialize};
+use std::convert::TryInto;
 
 entrypoint!(process_instruction);
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PriceData {
-    pub btc_price: u64,
-    pub eth_price: u64,
-    pub sol_price: u64,
+    pub btc_price: f64,
+    pub eth_price: f64,
+    pub sol_price: f64,
 }
 
 pub fn process_instruction(
@@ -31,14 +32,14 @@ pub fn process_instruction(
     }
 
     let mut price_data = PriceData {
-        btc_price: 0,
-        eth_price: 0,
-        sol_price: 0,
+        btc_price: 0.0,
+        eth_price: 0.0,
+        sol_price: 0.0,
     };
 
-    let btc_price = u64::from_le_bytes(instruction_data[0..8].try_into().unwrap());
-    let eth_price = u64::from_le_bytes(instruction_data[8..16].try_into().unwrap());
-    let sol_price = u64::from_le_bytes(instruction_data[16..24].try_into().unwrap());
+    let btc_price = f64::from_le_bytes(instruction_data[0..8].try_into().unwrap());
+    let eth_price = f64::from_le_bytes(instruction_data[8..16].try_into().unwrap());
+    let sol_price = f64::from_le_bytes(instruction_data[16..24].try_into().unwrap());
 
     price_data.btc_price = btc_price;
     price_data.eth_price = eth_price;
